Refactor file and test structure
Separate classes FlexTonalPalette and FlexCorePalette to own files
